+++ src/scenes/layouts/layouts.component.tsx
@@ -24,7 +24,7 @@ export const LayoutsScreen = (props): React.ReactElement => {
 
   return (
     <SafeAreaLayout style={styles.safeArea} insets="top">
-      <TopNavigation title="Cement App" accessoryLeft={renderDrawerAction} />
+      <TopNavigation title="Siva Tools" accessoryLeft={renderDrawerAction} />
       <Divider />
       <MenuGridList data={data} onItemPress={onItemPress} />
     </SafeAreaLayout>